ME2 - Main Config XIB behavior complete
For record: ME2 stands for Module Exercise number,
in this case, Module Exercise 2 (for second final module exercise)

Added:
- Fix uiview conflicts.
When stacks were child of a uiview, the objects inside were broken and couldn't be interacted with.
Made the stacks behave as the uiview in terms of constraints. (may be redundant to use a uiview to wrap the elements inside instead of just using the stack)

- byPoints TextField Delegates to accept numeric and backspace

- byRounds slider change by steps and show the string number

- Make selectedMode an enum to avoid miss spelling when writing. Two cases: .points and .rounds

- PlayButton visibility depending on selectedMode and if it is .points, PlayButton hides if at least one TextField is empty, shows if all TextFields (including Player's Name) have a value.
If it's .rounds, check only for Player's Name TextField.

To add:
- The next views: Info, Game and History
- Prepare segue to InfoView and GameView from Configuration